+++ lib/ui/profile/ProfileInfoPage.dart
@@ -4,16 +4,23 @@ import 'dart:typed_data';
 
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
+import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/ui/directory/DirectoryAccountsPage.dart';
 import 'package:illinois/ui/profile/ProfileInfoEditPage.dart';
 import 'package:illinois/ui/profile/ProfileInfoPreviewPage.dart';
 import 'package:illinois/ui/profile/ProfileInfoAndDirectoryPage.dart';
 import 'package:illinois/ui/directory/DirectoryWidgets.dart';
+import 'package:illinois/ui/profile/ProfileLoginPage.dart';
+import 'package:illinois/ui/settings/SettingsWidgets.dart';
+import 'package:illinois/ui/widgets/LinkButton.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:rokwire_plugin/service/auth2.dart';
 import 'package:rokwire_plugin/service/content.dart';
+import 'package:rokwire_plugin/service/groups.dart';
+import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
+import 'package:rokwire_plugin/service/social.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
@@ -22,11 +29,13 @@ class ProfileInfoPage extends StatefulWidget {
 
   final ProfileInfo contentType;
   final Map<String, dynamic>? params;
+  final bool showAccountCommands;
+  final void Function()? onStateChanged;
 
-  ProfileInfoPage({super.key, required this.contentType, this.params});
+  ProfileInfoPage({super.key, required this.contentType, this.params, this.showAccountCommands = false, this.onStateChanged});
 
   @override
-  State<StatefulWidget> createState() => _ProfileInfoPageState();
+  State<StatefulWidget> createState() => ProfileInfoPageState();
 
   bool? get editParam {
     dynamic edit = (params != null) ? params![editParamKey] : null;
@@ -34,7 +43,7 @@ class ProfileInfoPage extends StatefulWidget {
   }
 }
 
-class _ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileInfoPage> implements NotificationsListener {
+class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileInfoPage> implements NotificationsListener {
 
   Auth2UserProfile? _profile;
   Auth2UserPrivacy? _privacy;
@@ -44,6 +53,9 @@ class _ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileI
 
   bool _loading = false;
   bool _editing = false;
+  bool _preparingDeleteAccount = false;
+
+  bool get previewMode => !_loading && !_editing;
 
   @override
   void initState() {
@@ -66,6 +78,7 @@ class _ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileI
     if (name == DirectoryAccountsPage.notifyEditInfo) {
       setStateIfMounted((){
         _editing = true;
+        widget.onStateChanged?.call();
       });
     }
   }
@@ -87,15 +100,19 @@ class _ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileI
       );
     }
     else {
-      return ProfileInfoPreviewPage(
-        contentType: widget.contentType,
-        profile: _profile,
-        privacy: _privacy,
-        pronunciationAudioData: _pronunciationAudioData,
-        photoImageData: _photoImageData,
-        photoImageToken: _photoImageToken,
-        onEditInfo: _onEditInfo,
-      );
+      return Column(children: [
+        ProfileInfoPreviewPage(
+          contentType: widget.contentType,
+          profile: _profile,
+          privacy: _privacy,
+          pronunciationAudioData: _pronunciationAudioData,
+          photoImageData: _photoImageData,
+          photoImageToken: _photoImageToken,
+          onEditInfo: _onEditInfo,
+        ),
+        if (widget.showAccountCommands)
+          _accountCommands,
+      ],);
     }
   }
 
@@ -107,16 +124,100 @@ class _ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileI
     )
   );
 
+  Widget get _accountCommands => Column(children: [
+    Padding(padding: EdgeInsets.only(top: 8)),
+    _signOutButton,
+    _deleteAccountButton,
+    Padding(padding: EdgeInsets.only(top: 8)),
+  ],);
+
+  Widget get _signOutButton => LinkButton(
+    title: Localization().getStringEx('panel.profile.info.command.link.sign_out.text', 'Sign Out'),
+    textStyle: Styles().textStyles.getTextStyle('widget.button.title.small.underline'),
+    padding: EdgeInsets.only(top: 16, bottom: 8, left: 16, right: 16),
+    onTap: _onSignOut,
+  );
+
+  void _onSignOut() {
+    Analytics().logSelect(target: 'Sign Out');
+    showDialog<bool?>(context: context, builder: (context) => ProfilePromptLogoutWidget()).then((bool? result) {
+      if (result == true) {
+        Auth2().logout();
+      }
+    });
+  }
+
+  Widget get _deleteAccountButton => Stack(children: [
+    LinkButton(
+      title: AppTextUtils.appTitleString('panel.profile.info.command.link.delete_account.text', 'Delete My ${AppTextUtils.appTitleMacro} App Account'),
+      textStyle: Styles().textStyles.getTextStyle('widget.button.title.small.underline'),
+      padding: EdgeInsets.only(top: 8, bottom: 16, left: 16, right: 16),
+      onTap: _onDeleteAccount,
+    ),
+    if (_preparingDeleteAccount)
+      Positioned.fill(child:
+        Center(child:
+          SizedBox(width: 14, height: 14, child:
+            DirectoryProgressWidget()
+          )
+        )
+      )
+  ],);
+
+  void _onDeleteAccount() {
+    Analytics().logSelect(target: 'Delete Account');
+    if (!_preparingDeleteAccount) {
+      setState(() {
+        _preparingDeleteAccount = true;
+      });
+      Social().getUserPostsCount().then((int userPostCount) {
+        if (mounted) {
+          setState(() {
+            _preparingDeleteAccount = false;
+          });
+          final String groupsSwitchTitle = Localization().getStringEx('panel.settings.privacy_center.delete_account.contributions.delete.msg', 'Please delete all my contributions.');
+          SettingsDialog.show(context,
+              title: Localization().getStringEx("panel.settings.privacy_center.label.delete_message.title", "Delete your account?"),
+              message: [
+                TextSpan(text: Localization().getStringEx("panel.settings.privacy_center.label.delete_message.description1", "This will ")),
+                TextSpan(text: Localization().getStringEx("panel.settings.privacy_center.label.delete_message.description2", "Permanently "),style: Styles().textStyles.getTextStyle("widget.text.fat")),
+                TextSpan(text: Localization().getStringEx("panel.settings.privacy_center.label.delete_message.description3", "delete all of your information. You will not be able to retrieve your data after you have deleted it. Are you sure you want to continue?")),
+                if (0 < userPostCount)
+                  TextSpan(text:Localization().getStringEx("panel.settings.privacy_center.label.delete_message.description.groups", " You have contributed to Groups. Do you wish to delete all of those entries (posts, replies, reactions and events) or leave them for others to see.")),
+              ],
+              options: (0 < userPostCount) ? [groupsSwitchTitle] : null,
+              initialOptionsSelection: (0 < userPostCount) ?  [groupsSwitchTitle] : [],
+              continueTitle: Localization().getStringEx("panel.settings.privacy_center.button.forget_info.title","Forget My Information"),
+              onContinue: (List<String> selectedValues, OnContinueProgressController progressController) async {
+                Analytics().logAlert(text: "Remove My Information", selection: "Yes");
+                progressController(loading: true);
+                if (selectedValues.contains(groupsSwitchTitle)){
+                  Future.wait([Groups().deleteUserData(), Social().deleteUser()]);
+                }
+                await Auth2().deleteUser();
+                progressController(loading: false);
+                Navigator.pop(context);
+              },
+              longButtonTitle: true
+          );
+        }
+      });
+    }
+  }
+
   Future<void> _loadInitialContent() async {
     setState(() {
       _loading = true;
+      widget.onStateChanged?.call();
     });
+
     List<dynamic> results = await Future.wait([
       Auth2().loadUserProfile(),
       Auth2().loadUserPrivacy(),
       Content().loadUserPhoto(type: UserProfileImageType.medium),
       Content().loadUserNamePronunciation(),
     ]);
+
     if (mounted) {
       Auth2UserProfile? profile = JsonUtils.cast<Auth2UserProfile>(ListUtils.entry(results, 0));
       Auth2UserPrivacy? privacy = JsonUtils.cast<Auth2UserPrivacy>(ListUtils.entry(results, 1));
@@ -138,6 +239,7 @@ class _ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileI
         _photoImageData = photoResult?.imageData;
         _pronunciationAudioData = pronunciationResult?.audioData;
         _loading = false;
+        widget.onStateChanged?.call();
       });
     }
   }
@@ -186,6 +288,7 @@ class _ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileI
   void _onEditInfo() {
     setStateIfMounted(() {
       _editing = true;
+      widget.onStateChanged?.call();
     });
   }
 
@@ -216,6 +319,7 @@ class _ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileI
       }
 
       _editing = false;
+      widget.onStateChanged?.call();
     });
   }
 }